backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
requests
python-dotenv
scikit-learn
xgboost
shap
pandas
numpy
joblib
pdfplumber
python-docx
pytest
pytest-mock
httpx
